@@ -25,18 +25,14 @@ class AppFixtures extends Fixture
 
         TagFactory::createMany(100);
 
-        $questions = QuestionFactory::createMany(20, function () {
-            return [
-                'owner' => UserFactory::random(),
-            ];
-        });
+        $questions = QuestionFactory::createMany(20, fn() => [
+            'owner' => UserFactory::random(),
+        ]);
 
-        QuestionTagFactory::createMany(100, function () {
-            return [
-                'tag' => TagFactory::random(),
-                'question' => QuestionFactory::random(),
-            ];
-        });
+        QuestionTagFactory::createMany(100, fn() => [
+            'tag' => TagFactory::random(),
+            'question' => QuestionFactory::random(),
+        ]);
 
         QuestionFactory::new()
             ->unpublished()
@@ -44,16 +40,12 @@ class AppFixtures extends Fixture
             ->create()
         ;
 
-        AnswerFactory::createMany(100, function () use ($questions) {
-            return [
-                'question' => $questions[array_rand($questions)],
-            ];
-        });
-        AnswerFactory::new(function () use ($questions) {
-            return [
-                'question' => $questions[array_rand($questions)],
-            ];
-        })->needsApproval()->many(20)->create();
+        AnswerFactory::createMany(100, fn() => [
+            'question' => $questions[array_rand($questions)],
+        ]);
+        AnswerFactory::new(fn() => [
+            'question' => $questions[array_rand($questions)],
+        ])->needsApproval()->many(20)->create();
 
         $manager->flush();
     }
