@@ -11,11 +11,8 @@ use Symfony\Component\Security\Core\User\UserInterface;
 
 class QuestionVoter extends Voter
 {
-    private Security $security;
-
-    public function __construct(Security $security)
+    public function __construct(private Security $security)
     {
-        $this->security = $security;
     }
 
     protected function supports(string $attribute, $subject): bool
@@ -41,13 +38,10 @@ class QuestionVoter extends Voter
         if ($this->security->isGranted('ROLE_ADMIN')) {
             return true;
         }
-
         // ... (check conditions and return true to grant permission) ...
-        switch ($attribute) {
-            case 'EDIT':
-                return $user === $subject->getOwner();
-        }
-
-        return false;
+        return match ($attribute) {
+            'EDIT' => $user === $subject->getOwner(),
+            default => false,
+        };
     }
 }
