@@ -2,16 +2,11 @@
 
 declare(strict_types=1);
 
-use Rector\Symfony\Set\SymfonyLevelSetList;
-use Rector\Symfony\Set\SymfonySetList;
 use Rector\Config\RectorConfig;
+use Rector\Set\ValueObject\LevelSetList;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->symfonyContainerXml(__DIR__ . '/var/cache/dev/App_KernelDevDebugContainer.xml');
 
-$rectorConfig->sets([
-    SymfonyLevelSetList::UP_TO_SYMFONY_54,
-    SymfonySetList::SYMFONY_CODE_QUALITY,
-    SymfonySetList::SYMFONY_CONSTRUCTOR_INJECTION,
-]);
+$rectorConfig->sets([LevelSetList::UP_TO_PHP_80]);
 };
